Schema fixes (#318)
* rearrange code

* cleanup get_annotation_from_schema

* fix typo

* rename _schema to schema
